Removing the duplicate constant.js and application.js file
Removing the duplicate constant.js and application.js file
