+++ src-web/components/extensions/ConstantsExtension.js
@@ -1,0 +1,33 @@
+/*****************************************************************
+ *
+ * Copyright 2019 IBM Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ *****************************************************************/
+
+export const EXTENDED_RESOURCE_TYPES = {	
+    APPLICATION: { name: 'APPLICATION' },
+    JOB: { name: 'JOB' }
+}
+
+export const EXTENDED_CONFIG_CONSTANTS = {	
+    ACTION_MAP: 'action-map',
+    CMD_ACTIONS: 'cmd-actions',
+    INPUTS: 'inputs',
+    MENU_ITEM: 'menu-item',
+    REQUIRES_INPUT: 'requires-input',
+    URL_ACTIONS: 'url-actions',
+    ENABLEMENT_LABEL: 'enablement-label',
+    ENABLEMENT_ANNOTATION: 'enablement-annotation'
+}
